Add project id
When working with services like Vertex AI, you need to provide a project ID to ensure that your requests and operations are associated with the correct project and have the necessary permissions. The project ID serves as a unique identifier for your project within Google Cloud. So its compulsory to add it here!